feat: Add OnPopReturn widget to handle system back gesture and pop with value
This commit introduces a new widget, OnPopReturn, designed to manage the system back gesture and pop with a specified value. The OnPopReturn widget takes in a callback function to retrieve the value to be returned when the back gesture is detected. If the back gesture is not handled by the system, the widget will retrieve the value using the provided callback and pop with the value using MaterialNavigator.